# perf: load binaries in the same process