--- a/sources/commands/Enable.ts
+++ b/sources/commands/Enable.ts
@@ -5,6 +5,7 @@ import path                                                              from 'p
 import which                                                             from 'which';
 
 import {Context}                                                         from '../main';
+import * as nodeUtils                                                    from '../nodeUtils';
 import {isSupportedPackageManager, SupportedPackageManagerSetWithoutNpm} from '../types';
 
 export class EnableCommand extends Command<Context> {
@@ -51,7 +52,7 @@ export class EnableCommand extends Command<Context> {
     installDirectory = fs.realpathSync(installDirectory);
 
     // We use `eval` so that Webpack doesn't statically transform it.
-    const manifestPath = eval(`require`).resolve(`corepack/package.json`);
+    const manifestPath = nodeUtils.dynamicRequire.resolve(`corepack/package.json`);
 
     const distFolder = path.join(path.dirname(manifestPath), `dist`);
     if (!fs.existsSync(distFolder))