**What's the problem this PR addresses?**
<!-- Describe the rationale of your PR. -->
<!-- Link all issues that it closes. (Closes/Resolves #xxxx.) -->

While looking into improving Yarn's startup time, we've observed the fact that corepack adds extra overhead by spawning a new process to run the invoked binary.

**How did you fix it?**
<!-- A detailed description of your implementation. -->

This PR makes `corepack` load binaries in the same process, removing the `spawn` indirection and considerably improving the startup time:

<table>
<tr>
	<td>package-manager --version
	<td>Before
	<td>After
<tr>
	<td>yarn@3.2.0
	<td>183.1 ms
	<td>159.7 ms
<tr>
	<td>pnpm@6.29.1
	<td>92.0 ms
	<td>69.8 ms
<tr>
	<td>npm@8.1.0
	<td>160.8 ms
	<td>131.6 ms
</table>

The way it achieves this is by modifying `process.{argv,execArgv}` and `process.mainModule` (and the module cache to make sure that `require.main === module` keeps working) to trick the binary into thinking that it was spawned directly instead of having been loaded into the current process.

It might feel a bit magical at first, but we feel it's worth it.